# specs/006-tma-bridge/contracts/tma-bridge.openapi.yaml
openapi: 3.1.0
info:
  title: Tanzil TMA Bridge HTTP API
  version: 0.1.0
  description: HTTP contract for the Telegram Mini-App bridge.
servers:
  - url: /
paths:
  /api/health:
    get:
      summary: Health check for bridge readiness
      operationId: getHealth
      responses:
        "200":
          description: Bridge is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/session:
    post:
      summary: Exchange Telegram initData for a bridge session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Invalid or expired Telegram initData
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tasks:
    get:
      summary: List tasks visible to the authenticated Telegram user
      operationId: listTasks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Task list response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          description: Missing or invalid bridge session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new download task
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "202":
          description: Task accepted and forwarded to core
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandAck"
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid bridge session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tasks/{taskId}/actions:
    post:
      summary: Apply a control action to an existing task
      operationId: mutateTask
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskActionRequest"
      responses:
        "202":
          description: Action accepted and forwarded to core
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandAck"
        "400":
          description: Invalid action payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid bridge session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    SessionRequest:
      type: object
      required: [init_data]
      properties:
        init_data:
          type: string
          description: Raw Telegram Mini-App initData string as received from the client.
    SessionResponse:
      type: object
      required: [session_token, expires_at, user, theme, snapshot]
      properties:
        session_token:
          type: string
        expires_at:
          type: string
          format: date-time
        theme:
          type: string
          enum: [light, dark, system]
        user:
          $ref: "#/components/schemas/TelegramUser"
        snapshot:
          $ref: "#/components/schemas/TaskListResponse"
    TelegramUser:
      type: object
      required: [telegram_user_id]
      properties:
        telegram_user_id:
          type: integer
        username:
          type: string
        display_name:
          type: string
    TaskListResponse:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
    TaskSummary:
      type: object
      required: [task_id, status, created_at]
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        result:
          type: object
          additionalProperties: true
        error:
          type: string
    CreateTaskRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        label:
          type: string
    TaskActionRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [cancel]
    CommandAck:
      type: object
      required: [command_id, status]
      properties:
        command_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [accepted]
        task_id:
          type: string
          format: uuid
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string